+++ includes/minirt.h
@@ -52,7 +52,9 @@ void	ray_tracing(t_img *img, t_cam *cam, t_vw *vw, t_hit_lst *world);
 
 // minirt_utils.c
 void	init_mlx(t_rt *rt, t_img *img);
-void	init_world(t_cam *cam, t_hit_lst **world, char *filename);
 void	init_viewport(t_img *img, t_cam *camera, t_vw *viewport);
+
+// minirt_utils2.c
+void	init_world(t_cam *cam, t_hit_lst **world, char *filename);
 
 #endif
